Fix message-links-link-header being without a leading blank-line
Including the new-lines in the message-links-link-header meant
that removal of the blank line before it would no longer recognize the
header.

Now the newlines have been removed and searching for the link just
checks that the header is on it's own line without non-blank
characters before or after it.

Also add missing save-match-data to search.